move NEG stuff into pkg
