+++ pkg/networkendpointgroup/fakes.go
