+++ pkg/networkendpointgroup/interfaces.go
