+++ pkg/networkendpointgroup/syncer.go
